refactor: Remove `userid` from `Magmo3amodel`
The `userid` field has been removed from the `Magmo3amodel`. This simplifies the model by no longer storing the user's ID within the group's data, as it was redundant.

The changes affect the creation of new groups in `add_mogmo3a_cubit.dart` and the editing of existing groups in `add_magmo3a.dart`.