feat: Update color scheme and replace logo
This commit introduces a new color palette by updating the hex codes for `darkGrey`, `ligthGreen`, and `green` in `lib/colors_app.dart`, and removes the unused `blue` color.

The app's logo has been replaced with a new asset (`studenizer_logo_2.png`), which is now displayed on the home page's day tabs. The app icon in the `AppBar` and the text in the `Magmo3aWidget` have also been updated to use the new color scheme.